backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.30
aiosqlite==0.20.0
Pillow==10.3.0
numpy==1.26.4
torch==2.2.2
torchvision==0.17.2
opencv-python-headless==4.9.0.80
matplotlib==3.8.4
scipy==1.13.0
scikit-learn==1.4.2
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
fpdf2==2.7.9
